undebug

+++ testkit/src/mill/testkit/IntegrationTesterBase.scala
@@ -59,11 +59,13 @@ trait IntegrationTesterBase {
    * Remove any ID files to try and force them to exit
    */
   def removeProcessIdFile(): Unit = {
-    val serverPath0 =
-      os.Path(out, workspacePath) / (if (clientServerMode) millServer else millNoServer)
+    val outDir = os.Path(out, workspacePath)
+    if (os.exists(outDir)) {
+      val serverPath0 = outDir / (if (clientServerMode) millServer else millNoServer)
 
-    for (serverPath <- os.list.stream(serverPath0)) os.remove(serverPath / processId)
+      for (serverPath <- os.list.stream(serverPath0)) os.remove(serverPath / processId)
 
-    Thread.sleep(500) // give a moment for the server to notice the file is gone and exit
+      Thread.sleep(500) // give a moment for the server to notice the file is gone and exit
+    }
   }
 }
